--- src/components/register/register.css
@@ -1,67 +1,0 @@
-.register-container {
-    width: 100%;
-    max-width: 1000px;
-    margin: auto;
-    padding: 30px;
-    border: 1px solid #ddd;
-    border-radius: 10px;
-    background-color: #ffffff;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-    text-align: left;
-  }
-  
-  .register-container h2 {
-    text-align: center;
-    margin-bottom: 20px;
-    color: #333;
-    font-size: 24px;
-  }
-  
-  .register-container h3 {
-    font-size: 14px;
-    text-align: left;
-    margin-bottom: 10px;
-    color: #555;
-  }
-  
-  .register-container form {
-    display: flex;
-    flex-direction: column;
-    color: #333;
-  }
-  
-  .register-container label {
-    margin-bottom: 5px;
-    font-weight: bold;
-    color: #555;
-  }
-  
-  .register-container input {
-    background-color: #f5f5f5;
-    color: #333;
-    padding: 12px;
-    margin-bottom: 15px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    transition: border-color 0.3s;
-  }
-  
-  .register-container input:focus {
-    border-color: #007bff;
-    outline: none;
-  }
-  
-  .register-container button {
-    margin-top: 20px;
-    padding: 12px;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: background-color 0.3s;
-  }
-  
-  .register-container button:hover {
-    background-color: #0056b3;
-  }
